+++ simple-stack/src/main/java/com/zhuinden/simplestack/BackstackDelegate.java
@@ -542,6 +542,20 @@ public class BackstackDelegate {
     }
 
     /**
+     * Returns a list of the scopes accessible from the given key.
+     *
+     * The order of the scopes is that the 0th index is the current scope (if available), followed by parents.
+     *
+     * @param key        the key
+     * @param lookupMode determine what type of parents are checked during the lookup
+     * @return the list of scope tags
+     */
+    @NonNull
+    public List<String> findScopesForKey(@NonNull Object key, @NonNull ScopeLookupMode lookupMode) {
+        return getManager().findScopesForKey(key, lookupMode);
+    }
+
+    /**
      * Attempts to look-up the service in the scopes accessible from the provided scope tag.
      * Returns whether the service exists in any of these scopes.
      *